Moved file code to common directory
