+++ backend/package.json
@@ -33,6 +33,7 @@
     "testEnvironment": "node"
   },
   "dependencies": {
+    "cc-map-editor-common": "../common",
     "async": "^2.6.0",
     "body-parser": "^1.18.2",
     "cors": "^2.8.4",
