+++ common/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "cc-map-editor-common",
+  "version": "0.0.1",
+  "description": "",
+  "main": "main.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "build": "tsc",
+    "watch": "tsc -w"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/CCDirectLink/crosscode-map-editor.git"
+  },
+  "author": "CCDirectLink",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/CCDirectLink/crosscode-map-editor/issues"
+  },
+  "homepage": "https://github.com/CCDirectLink/crosscode-map-editor#readme",
+  "devDependencies": {
+    "tslint": "^5.16.0",
+    "typescript": "^3.4.5"
+  }
+}
